Consolidate icon styles
