+++ packages/buttons/src/styled/StyledIconButton.ts
@@ -6,10 +6,8 @@
  */
 
 import styled, { css, ThemeProps, DefaultTheme } from 'styled-components';
-import { math } from 'polished';
 import { retrieveComponentStyles, getColor, DEFAULT_THEME } from '@zendeskgarden/react-theming';
 import { StyledButton, getHeight, IStyledButtonProps } from './StyledButton';
-import { StyledIcon } from './StyledIcon';
 
 const COMPONENT_ID = 'buttons.icon_button';
 
@@ -45,18 +43,6 @@ const iconButtonStyles = (props: IStyledButtonProps & ThemeProps<DefaultTheme>) 
   `;
 };
 
-const iconStyles = (props: IStyledButtonProps & ThemeProps<DefaultTheme>) => {
-  const size =
-    props.size === 'large'
-      ? math(`${props.theme.space.base * 6} * 1px`)
-      : math(`${props.theme.space.base * 4} * 1px`);
-
-  return css`
-    width: ${size};
-    height: ${size};
-  `;
-};
-
 /**
  * Accepts all `<button>` props
  */
@@ -65,10 +51,6 @@ export const StyledIconButton = styled(StyledButton).attrs(() => ({
   'data-garden-version': PACKAGE_VERSION
 }))`
   ${props => iconButtonStyles(props)};
-
-  ${StyledIcon} {
-    ${props => iconStyles(props)}
-  }
 
   ${props => retrieveComponentStyles(COMPONENT_ID, props)};
 `;
